+++ src/pages/administration/students/index.unit.js
@@ -56,7 +56,7 @@ describe("students/index", () => {
 			},
 			users: {
 				actions: {
-					handleUsers: jest.fn(),
+					findStudents: jest.fn(),
 					deleteUsers: deleteUsersStub,
 					getQrRegistrationLinks: jest.fn(),
 					sendRegistrationLink: jest.fn(),
@@ -168,14 +168,14 @@ describe("students/index", () => {
 		});
 	});
 
-	it("should dispatch the 'handleUsers action on load'", async () => {
+	it("should dispatch the 'findStudents action on load'", async () => {
 		mount(StudentPage, {
 			...createComponentMocks({
 				i18n: true,
 				store: mockStore,
 			}),
 		});
-		expect(mockStore.users.actions.handleUsers).toHaveBeenCalled();
+		expect(mockStore.users.actions.findStudents).toHaveBeenCalled();
 	});
 
 	it("should emit the 'delete' action when deleting a user", async () => {
@@ -495,7 +495,7 @@ describe("students/index", () => {
 		//run all existing timers
 		await jest.runAllTimers();
 		//reset the mock call stack
-		mockStore.users.actions.handleUsers.mockClear();
+		mockStore.users.actions.findStudents.mockClear();
 		mockStore.uiState.mutations.set.mockClear();
 
 		const searchBarInput = wrapper.find(`input[data-testid="searchbar"]`);
@@ -508,7 +508,7 @@ describe("students/index", () => {
 		//run new timer from updating the value
 		await jest.runAllTimers();
 
-		expect(mockStore.users.actions.handleUsers).toHaveBeenCalled();
+		expect(mockStore.users.actions.findStudents).toHaveBeenCalled();
 	});
 
 	// currently disabled, will be reactivated when the new components are in use