+++ CHANGELOG.md
@@ -8,6 +8,19 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ## Unreleased
 
+### Added
+
+- SC-9090 - implemented a loading state in the service-template
+
+### Changed
+
+- SC-9077 - refactored store actions returning data
+- SC-9126 - changed the way the stores state is accessed; `mapState` and `store.state` were substituted for `mapGetters` and `store.getters` respectively; and changed the naming of the template and several other stores getters
+
+### Fixed
+
+- SC-9131 fixed the consent table loading the wrong list of users and refactored the consent page
+- SC-9106 - Remove the possibility to change an action's endpoint with a param from the component
 - SC-9093 - Load default language and timezone from server
 - SC-9158 - fixed launch cofig to debug a single test
 
@@ -54,7 +67,6 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 - SC-8156 - added Spanish as an available language
 - SC-9004 - Sync env variables between backend and frontend
 - OPS-1499 - added feature Branch to CI Pipeline for team based HPI Schul-Cloud instances
-- SC-9090 - implemented a loading state in the service-template
 
 ### Changed
 
@@ -65,8 +77,6 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 - SC-7944 - changes field for Lern-Store content id and adds a link to collection
 - OPS-1508 - Updated limits for cpu and ram to the docker compose files ( 20.11.2020 )
 - SC-9004 - Change ci build
-- SC-9077 - refactored store actions returning data
-- SC-9126 - changed the way the stores state is accessed; `mapState` and `store.state` were substituted for `mapGetters` and `store.getters` respectively; and changed the naming of the template and several other stores getters
 
 ### Fixed